# dodoist-docs/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Dodoist API
  version: 1.0.0
  description: Task and project management API combining personal todos with Scrum/Kanban
    workflows.
paths:
...
  /api/projects/{id}/snapshots/:
    get:
      operationId: projects_snapshots_retrieve
      description: GET /api/projects/<pk>/snapshots/
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - projects
      responses:
        '200':
          description: No response body
...